requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
pydantic>=2.7.0
tenacity>=9.0.0
python-dotenv>=1.0.1
pyttsx3>=2.90
